+++ CHANGELOG.md
@@ -1,18 +1,31 @@
 ## 0.0.1
 
-* TODO: Initial version.
+- TODO: Initial version.
 
 ## 0.0.2
-* Bug fix for Android save 
+
+- Bug fix for Android save
 
 ## 0.0.3
-* Bug fix and add utility functions  
+
+- Bug fix and add utility functions
 
 ## 0.0.4
-* Bug fix
+
+- Bug fix
 
 ## 0.0.5
-* Support to read image original bytes
+
+- Support to read image original bytes
 
 ## 0.0.7
-* iOS: Fix read byte data from URL
+
+- iOS: Fix read byte data from URL
+
+## 0.0.8
+
+- Android: Upgrade to Android SDK 35
+
+## 1.0.0
+
+- First major version